CHROMIUM PORT OF HTTPZ
Here is a brief description of the things I did to make this port possible. Added PNG file icon sizes to the assets folder to be used in the browser toolbar. Edited the html pages to include the browser-polyfill script and mobile viewport meta tag. Declared a fuction name in the content script js file.